perf(ui): consolidate isolated chat pane lifecycle tests
