+++ test/vitest/vitest.ui-isolated-paths.mjs
@@ -19,7 +19,6 @@ export const uiIsolatedTestFiles = [
   "ui/src/pages/chat/chat-pane-lifecycle.test.ts",
   "ui/src/pages/chat/chat-pane-pull-requests.test.ts",
   "ui/src/pages/chat/chat-pane-retained-presentation.test.ts",
-  "ui/src/pages/chat/chat-pane.message-cut.test.ts",
   "ui/src/pages/chat/chat-pane.read-marker.test.ts",
   "ui/src/pages/chat/chat-pane.session-discussion.test.ts",
   "ui/src/pages/chat/chat-pane.test.ts",
