+++ ui/src/pages/chat/chat-pane.test-support.ts
@@ -55,6 +55,7 @@ export type TestChatPane = HTMLElement & {
   connectionGeneration: number;
   catalogLoadGeneration: number;
   continueCatalogSession: (key: CatalogSessionKey) => Promise<void>;
+  forkFromMessage: (entryId: string) => Promise<void>;
   createSession: () => Promise<boolean>;
   recoverSession: () => Promise<boolean>;
   restartRecoveryComposerBanner: () =>
